Fix get jumping note spawn data causing crashes, CJD bug? (#5)
* Fix GetJumpingNoteSpawnData SEGV_MAPERR crash

* Make the stack with the double flow happy by making the stack code easier to read

* Notes are actually flipping now

+++ src/Hooks/BeatmapObjectsInTimeRowProcessor.cpp
@@ -15,7 +15,7 @@ MAKE_HOOK_MATCH(BeatmapObjectsInTimeRowProcessor_ProcessAllNotesInTimeRow,
     BeatmapObjectsInTimeRowProcessor_ProcessAllNotesInTimeRow(self, notes);
 
     auto *customNotes = reinterpret_cast<List<CustomNoteData *> *>(notes);
-    std::unordered_map<int, std::vector<CustomNoteData *>> notesInColumn;
+    std::unordered_map<float, std::vector<CustomNoteData *>> notesInColumn;
     for (int i = 0; i < customNotes->get_Count(); i++) {
         CustomNoteData *noteData = customNotes->items->values[i];
